requirements.txt
aniso8601==8.0.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
Pillow==6.2.1
pymongo==3.10.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-file==1.4.3
six==1.13.0
soupsieve==1.9.5
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.7
Werkzeug==0.16.0
